Refactor crop advanced panel controls
